refactor: move post stories components into features with stubs
