+++ src/pages/ProfilePage.tsx
@@ -6,7 +6,7 @@ import { useAuth } from '../context/AuthContext';
 import { subscribeToUserProfile, type UserProfileRead } from '../lib/firestore';
 import EditProfileModal from '../components/EditProfileModal';
 import CollectionsPanel from '../components/CollectionsPanel';
-import StoriesWidget from '../components/StoriesWidget';
+import StoriesWidget from '@/features/posts/components/StoriesWidget';
 import ProfilePostsGrid from '../components/ProfilePostsGrid';
 import ContributionsSection from '../components/ContributionsSection';
 import { CATEGORIES } from '../data/mockData';
